[Feature] Add spec-compliant API: napi_call_function_spec_compliant
Since the current behavior of PrimJS's `napi_call_function` API is inconsistent with the standard in JSC backend, the NAPI Adapter requires a new API to achieve alignment with the standard. To avoid breaking changes for existing users, the new `napi_call_function_spec_compliant` API is added.

issue: m-6565624125