[Feature] Add spec-compliant API: napi_wrap_spec_compliant & napi_unwrap_spec_compliant & napi_remove_wrap_spec_compliant
Since the current behavior of PrimJS's `napi_wrap`, `napi_unwrap`, `napi_remove_wrap` API is inconsistent with the standard in JSC backend, the NAPI Adapter requires a new API to achieve alignment with the standard. To avoid breaking changes for existing users, the new spec-compliant API is added.

issue: m-6565624125